% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{daily}
\alias{daily}
\alias{weekly}
\alias{monthly}
\alias{daily_basic}
\alias{adj_factor}
\alias{moneyflow}
\alias{fund_daily}
\alias{opt_daily}
\alias{index_daily}
\title{Daily stock market data.}
\usage{
daily(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

weekly(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

monthly(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

daily_basic(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

adj_factor(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

moneyflow(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

fund_daily(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

opt_daily(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))

index_daily(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"))
}
\arguments{
\item{ts_code}{OPTIONAL, Tushare equity code.}

\item{trade_date}{OPTIONAL, trading date.}

\item{start_date}{OPTIONAL, start date of returned data.}

\item{end_date}{OPTIONAL, end date of returned date.}

\item{date_format}{How to cast return datetime columns? Default casts values to POSIXct objects.}
}
\value{
a data.frame/data.table
}
\description{
Please refer to online document for details.
}
\examples{
\dontrun{
dt <- daily(ts_code = "000001.sz")
dt <- fut_daily(trade_date = "20190301")
}
}
