% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{fut_basic}
\alias{fut_basic}
\title{Get basic info of futures contracts.}
\usage{
fut_basic(exchange, fut_type = "", date_format = c("POSIXct", "Date",
  "char"), ...)
}
\arguments{
\item{exchange}{Future exchagne centre.}

\item{fut_type}{Type of futures.}

\item{date_format}{How to cast datetime format.}

\item{...}{futher arguments passed}
}
\value{
data.frame/data.table
}
\description{
Get basic info of futures contracts.
}
\examples{
\dontrun{
fut_basic(exchange="DCE", fut_type="1")
}
}
