% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{fut_holding}
\alias{fut_holding}
\alias{fut_wsr}
\alias{fut_daily}
\alias{fut_settle}
\title{Futures market data.}
\usage{
fut_holding(trade_date = "", symbol = "", start_date = "",
  end_date = "", exchange = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fut_wsr(trade_date = "", symbol = "", start_date = "",
  end_date = "", exchange = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fut_daily(ts_code = "", trade_date = "", start_date = "",
  end_date = "", exchange = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fut_settle(ts_code = "", trade_date = "", start_date = "",
  end_date = "", exchange = "", date_format = c("POSIXct", "Date",
  "char"), ...)
}
\arguments{
\item{trade_date}{OPTIONAL, trading date.}

\item{symbol}{OPTIONAL, futures symbol.}

\item{start_date}{OPTIONAL, start date of returned data.}

\item{end_date}{OPTIONAL, end date of returned data.}

\item{exchange}{OPTIONAL, Future exchange centre.}

\item{date_format}{How to cast datetime format.}

\item{...}{futher argument passed}

\item{ts_code}{OPTIONAL, Tushare equity code.}
}
\value{
data.frame/data.table
}
\description{
Futures market data.
}
\examples{
\dontrun{
fut_holding(trade_date='20181113', symbol='C', exchange='DCE')
fut_wsr(trade_date='20181113', symbol='ZN')
}
}
