% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynapi.R
\name{GetAPI}
\alias{GetAPI}
\title{Get a Tushare API object.}
\usage{
GetAPI(token = GetToken())
}
\arguments{
\item{token}{API token.}
}
\value{
a tushare_api object.
}
\description{
Get a Tushare API object.
}
