% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{income}
\alias{income}
\alias{income_vip}
\alias{balancesheet}
\alias{balancesheet_vip}
\alias{cashflow}
\alias{cashflow_vip}
\alias{forecast}
\alias{forecast_vip}
\alias{express}
\alias{express_vip}
\alias{fina_indicator}
\alias{fina_indicator_vip}
\alias{fina_audit}
\alias{fina_mainbz}
\alias{fina_mainbz_vip}
\title{Consolidated financial reports.}
\usage{
income(ts_code = "", ann_date = "", start_date = "", end_date = "",
  period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

income_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

balancesheet(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

balancesheet_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

cashflow(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

cashflow_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", report_type = "", comp_type = "",
  date_format = c("POSIXct", "Date", "char"), ...)

forecast(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

forecast_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

express(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

express_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fina_indicator(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fina_indicator_vip(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fina_audit(ts_code = "", ann_date = "", start_date = "",
  end_date = "", period = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fina_mainbz(ts_code = "", period = "", type = c("", "P", "D"),
  start_date = "", end_date = "", date_format = c("POSIXct", "Date",
  "char"), ...)

fina_mainbz_vip(ts_code = "", period = "", type = c("", "P", "D"),
  start_date = "", end_date = "", date_format = c("POSIXct", "Date",
  "char"), ...)
}
\arguments{
\item{ts_code}{OPTIONAL, Tushare equity code.}

\item{ann_date}{OPTIONAL, annoucement date.}

\item{start_date}{OPTIONAL, report start date.}

\item{end_date}{OPTIONAL, report end date.}

\item{period}{OPTIONAL, report period.}

\item{report_type}{OPTIONAL, report type. Please refer to online API document.}

\item{comp_type}{OPTIONAL, report company type. Please refer to online API document.}

\item{date_format}{How to cast return datetime columns? Default casts values to POSIXct objects.}

\item{...}{further arguments passed}

\item{type}{OPTIONAL, business type.}
}
\value{
a data.frame/data.table
}
\description{
Consolidated financial reports.
}
\examples{
\dontrun{
dt <- income(ts_code='600000.SH', start_date='20180101', end_date='20180730')
}
}
