% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{market_eod}
\alias{market_eod}
\title{Get end-of-day market data.}
\usage{
market_eod(ts_code = "", trade_date = "", start_date = "",
  end_date = "", date_format = c("POSIXct", "Date", "char"),
  api = c("daily", "weekly", "monthly", "adj_factor", "daily_basic",
  "moneyflow", "stk_limit", "index_daily", "index_weekly", "index_monthly",
  "index_weight", "index_dailybasic", "fund_daily", "fut_daily",
  "opt_daily"))
}
\arguments{
\item{ts_code}{Tushare equity code}

\item{trade_date}{Trading date}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{date_format}{How to cast datetime columns}

\item{api}{actual API function}
}
\value{
data.table
}
\description{
Get end-of-day market data.
}
