% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{POSTRetry}
\alias{POSTRetry}
\title{Wrapped httr::POST with retry. Rethrow errors as warnings.}
\usage{
POSTRetry(..., retry, sleep = 0, warn = TRUE)
}
\arguments{
\item{...}{passed to httr::POST()}

\item{retry}{max retries}

\item{sleep}{seconds to sleep if error occurs}

\item{warn}{whether to throw warnings if error occurs}
}
\value{
reponse
}
\description{
Wrapped httr::POST with retry. Rethrow errors as warnings.
}
