% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynapi.R
\name{print.tushare_api}
\alias{print.tushare_api}
\title{Print Values}
\usage{
\method{print}{tushare_api}(x, ...)
}
\arguments{
\item{x}{A tushare_api object}

\item{...}{not used}
}
\value{
x, invisibly
}
\description{
Print Values
}
