% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{SetToken}
\alias{SetToken}
\title{Set Tushare Pro API token}
\usage{
SetToken(token)
}
\arguments{
\item{token}{a character vector}
}
\value{
token itself, invisibly.
}
\description{
Set Tushare Pro API token
}
\examples{
SetToken("YOUR_API_TOKEN")
}
