% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{stock_basic}
\alias{stock_basic}
\alias{trade_cal}
\alias{hs_const}
\alias{namechange}
\alias{stock_company}
\alias{new_share}
\title{Get basic stock market information}
\usage{
stock_basic(is_hs = c("", "N", "H", "S"), list_status = c("", "L", "D",
  "P"), exchange = c("", "SSE", "SZSE", "HKEX"),
  date_format = c("POSIXct", "Date", "char"))

trade_cal(exchange = c("", "SSE", "SZSE"), start_date = "",
  end_date = "", is_open, date_format = c("POSIXct", "Date", "char"))

hs_const(hs_type = c("SH", "SZ"), is_new = c("1", "0"),
  date_format = c("POSIXct", "Date", "char"))

namechange(ts_code = "", start_date = "", end_date = "",
  date_format = c("POSIXct", "Date", "char"))

stock_company(exchange = c("SSE", "SZSE"), date_format = c("POSIXct",
  "Date", "char"))

new_share(start_date = "", end_date = "", date_format = c("POSIXct",
  "Date", "char"))
}
\arguments{
\item{is_hs}{Flag for Shenzhen/Shanghai-Hong Kong Stock Connect}

\item{list_status}{List status}

\item{exchange}{Exchange market}

\item{date_format}{How to cast datetime format.}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{is_open}{Wheter the market is open}

\item{hs_type}{Type of Shenzhen/Shanghai-Hong Kong Stock Connect}

\item{is_new}{Whether the stock is new}

\item{ts_code}{Tushare equity code}
}
\value{
data.frame/data.table
}
\description{
Get basic stock market information
}
\examples{
\dontrun{
stock_basic()
}
}
