% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{suspend}
\alias{suspend}
\title{Suspend information}
\usage{
suspend(ts_code = "", suspend_date = "", resume_date = "",
  date_format = c("POSIXct", "Date", "char"), ...)
}
\arguments{
\item{ts_code}{OPTIONAL, Tushare equity code.}

\item{suspend_date}{OPTIONAL, suspend date in YYYYmmdd format or Date/POSIXct object.}

\item{resume_date}{OPTIONAL, resume date in YYYYmmdd format or Date/POSIXct object.}

\item{date_format}{How to cast return datetime columns? Default casts values to POSIXct objects.}

\item{...}{futher arguments passed}
}
\value{
a data.frame/data.table
}
\description{
Suspend information
}
\examples{
\dontrun{
dt <- suspend(ts_code = "000001.sz")
}
}
