% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{TimedPOST}
\alias{TimedPOST}
\title{Wrapped httr::POST with retry for timeout error.}
\usage{
TimedPOST(..., retry)
}
\arguments{
\item{...}{passed to httr::POST()}

\item{retry}{max retries}
}
\value{
response
}
\description{
Notes on TimedPOST: non-timeout errors are rethrown
}
