% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000tusharer.R
\docType{package}
\name{TushareR-package}
\alias{TushareR}
\alias{TushareR-package}
\title{TushareR: R Interface To Tushare Pro API}
\description{
R interface to Tushare Pro API.
}
\details{
TushareR is a wrapper for Tushare Pro API \url{https://tushare.pro}. Please
refer to the online document for details of the APIs.

Prior to using TushareR, you need to set your API token by calling
SetToken("YOUR_API_TOKEN").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imlijunda/TushareR}
  \item Report bugs at \url{https://github.com/imlijunda/TushareR/issues}
}

}
