% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{TusRequest}
\alias{TusRequest}
\title{Make raw request to Tushare Pro API}
\usage{
TusRequest(api_name, ...)
}
\arguments{
\item{api_name}{name of API function, please refer to online document for more information.}

\item{...}{passed to API function.}
}
\value{
data.frame/data.table
}
\description{
Make raw request to Tushare Pro API
}
\examples{
\dontrun{
top10 <- TusRequest("top10_holders", ts_code = "000001.SZ")
}
}
