% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{SetToken}
\alias{SetToken}
\alias{GetToken}
\title{Set/Get Tushare Pro API token}
\usage{
SetToken(token)

GetToken()
}
\arguments{
\item{token}{a character vector}
}
\value{
token itself, invisibly.
}
\description{
Set/Get Tushare Pro API token
}
\examples{
SetToken("YOUR_API_TOKEN")
GetToken()
}
